+++ source/tests/test_string_manipulation.py
@@ -49,3 +49,32 @@ def test_convert_list_values_to_string():
     list_of_strings = string_manipulation.convert_list_values_to_string(list_of_numbers)
     for string in list_of_strings:
         assert isinstance(string, str)
+
+
+def test_convert_string_to_list_default_separator():
+    separator = ','
+    value = "a, b"
+    list_1 = value if separator not in value else \
+        string_manipulation.convert_string_to_list(value, separator)
+    assert isinstance(list_1, list)
+    assert list_1[0] == 'a'
+    assert list_1[1] == 'b'
+
+
+def test_convert_string_to_list_no_separator():
+    separator = ','
+    value = "a"
+    string = value if separator not in value else \
+        string_manipulation.convert_string_to_list(value, separator)
+    assert isinstance(string, str)
+    assert string == 'a'
+
+
+def test_convert_string_to_list_custom_separator():
+    separator = ';'
+    value = "a; b"
+    list_1 = list_1 = value if separator not in value else \
+        string_manipulation.convert_string_to_list(value, separator)
+    assert isinstance(list_1, list)
+    assert list_1[0] == 'a'
+    assert list_1[1] == 'b'